# src/main/java/my/project/user_note/entity/note/Note.java
package my.project.user_note.entity.note;

import jakarta.persistence.Entity;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.ToString;

import java.util.UUID;

@Entity
@Getter
@Setter
@ToString
@NoArgsConstructor
@Table(name = "NOTE")
public class Note {
    @Id
    private UUID id;
    private String text;
}